Doing the thing directly instead of calling git.exe to do it
By calling git, not everyone has git installed in the same folder as AoTTG2 proj so in many case that would throw an exception or an error as the programm wouldn't be found so it wouldn't update the version as it was supposed to.
All the system already did everything else, it was just this little thing to create some issue, as now it check directly the git file inside the repository which mean that will be in the same position for every branch and no matter the version of git or how/where you installed it in your machine.
As this change the version of the game it will also separate the multiplayer gameplays by branch.
Only client from the same exact branch can play with each other on multi.