Base Implementation
A basic implementation of a translation system based on the use of .po files, it need further development for the automation of the unityeditor script to create a file which hold the available translations, and also some working on other things. Is just a first step on the development. Everything will be handled by the Language class whilst the loading of the dictionary holding the translation and the loading of the list of available translation will be held by the other classes, i also added a 4th class, to replace the unity.UI.Text class in order to automatically translate also UI elements without the need of manually do it. I'll add a script to it later to work on editor so that it would pop up in the UI drop down menù together with the Unity UI as translatedText or internationalizedText or similar