Refactor Service.Player.HeroHit event and add Service.Player.HeroKll event.
Prior to this a HeroKillEvent would've been triggered whenever you another hero with yours blades, EVEN if they were on the same team. The code was also quite ambiguous; the event is called HeroHit but in reality it's only triggered when an enemy hero is killed.

So this change differentiates between hitting a hero and killing a hero by adding a new HeroKill event.