Fix GUEST kills GUEST in multiplayer.
This issue occured for any attacks that use HitEntity() (i.e. the double fist combo attack). This fix works because if viewID = -1 then Hero.die() will use TItan.name as the killer's name. If not, then Hero.die() will use the Photon ID of the killer. Thus, for player titans we will not use -1 so it shows the name of the player playing the player titan.