Folders For Prefabs Added
I have added folders in order to keep the prefabs organized for use in maps and scenes. Each "Category" has a material folder to be used for everything under that "Category." (Make sure to keep this organized with proper material names) This needs to be checked over before I begin placing all prefabs into these folders.