Add logic for reloading Characters List in spawn menu
Copied some logic from SpawnMenu awake function into OnEnable, so it reloads the Characters dropdown whenever a lobby is joined, so default character is selected as it should be. Previously upon entering new lobby the outfit and build for the characters was reloaded and read from the first character on the list, despite selected character never being reset, which resulted in desynchronization between selected character, and it's possible outfit etc.