SpawnMenu loads outfit of selected instead of first
Now the list of outfits and builds defaults for the selected character.
Previously SpawnMenu, upon being enabled, meaning when entering a new lobby would set the possible outfit options to those of the first character on the list, instead of the one selected. Right now the selected character happens to be the first one on the list, but there might be some logic introduced later on, that would change that, which would again result in desynchonization between defaultly selected character and it's outfits.