fixed the barrier blocks and placed some blocks in the corners for orientation
The barrier blocks were placed lazy and therefore ther was some loss of terrain and I now made it somewhat a perfect cube. And I placed 4 blocks in each corner of the map so I have orientation when I later copy the map and paste it on the other site.