Animated Server Selection and setted scalable main menu to work with different screen ratio
include the addition to ServerSelector class in order to animate the panel through callback. Introduce the mapbuttonanimation class to animate the various region selection as a child of the buttonanimation class, used just to animate the buttons in the menu, also remove some duplicates and some obsolete files and animation of the manu no more used, also clean even more the scene aottg2.