Fixed Scene For UltraWide Screen
Added function to scale depth of field with the screen size and added scale for screen ratio and field of view to have a more consistent view through the various possible screen size and aspect ratios. Added some details and extended the scene for ultra wide screen (32:9)
Also initalized localization for ServerSelector screen.