No camera movement when locking
To stop the camera from flipping, I simply stopped moving the camera if the player is currently locked into a titan.
In addition, this makes locking a better experience because the camera is no longer shaking on mouse input.
Note that I identified the culprit, line 580 in the modified file, but didn't modify it, as I'm not sure what it does.