Fix incorrect offset when shooting multiple hooks
I'm assuming !single refers to shooting both hooks. I haven't tested this, but I can see that I made a mistake when refactoring this part of the code. Both hooks probably got shot to the left in the previous build.
This also gets rid of the code smell that was picked up by SonarCloud.