Update to Crossfade
Made it so crossfading animations works more smoothly with the new state machine (can also use longer fade times to make the transitions less instantaneous). Added a new function to the Interface that is called in the Hero Movement State class that ensures the right animation is playing all the time. Also removes it from all the child classes that inherit from Hero Movement cleaning it up a bit